% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{plot.tebounds}
\alias{plot.tebounds}
\title{Plot Treatment Effect Bounds}
\usage{
\method{plot}{tebounds}(x)
}
\arguments{
\item{x}{An object of the class "tebounds"}
}
\value{
A list of plots of treatmetn effect bounds
}
\description{
Plot Treatment Effect Bounds
}
