% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.tebounds}
\alias{summary.tebounds}
\title{Summarize tebound object}
\usage{
\method{summary}{tebounds}(x, ndigits = 3)
}
\arguments{
\item{x}{An object of the class "tebounds".}

\item{ndigits}{Number of digits to round to.}
}
\value{
A summary of treatment effect bounds
}
\description{
Summarize tebound object
}
\keyword{internal}
