% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tebounds.R
\name{tebounds}
\alias{tebounds}
\title{tebounds}
\usage{
tebounds(
  data,
  formula,
  weights,
  subset,
  erates = c(0, 0.05, 0.1, 0.25),
  np = 500,
  bootstrap = F,
  boot.reps = 100,
  CI.level = 95,
  im = F,
  k = 100,
  ncells = 5
)
}
\arguments{
\item{data}{A data frame or survey design object containing a binary outcome (Y) and a binary treatment (D).
The data can optionally include a monotone instrumental variable (Z) and a vector of weights.}

\item{formula}{A formula indicating the outcome, treatment, and instrument (if any). Instruments are specified using the "|"  at the end of the formula (e.g. Y~D|Z).}

\item{weights}{Data frame column indicating weight variable, provided without quotation marks.  If survey design with weights is supplied, then this argument is not needed.}

\item{subset}{Conditions to subset data.}

\item{erates}{Numeric vector, set to (0, 0.5, 0.1, 0.25) by default.  Vector specifying assumed rates of measurement error.}

\item{np}{Integer, set to 500 by default, specifying number of intervals for grid search in arbitrary measurement-error model}

\item{bootstrap}{Boolean, set to FALSE by default. Set to TRUE to estimate confidence intervals (CIs) via bootstrap}

\item{boot.reps}{Integer, set to 100 by default, specifying number of bootstrap replications for confidence interval estimation.}

\item{CI.level}{Integer from 0 to 100, set to 95 by default.  Level at which to calculate confidence intervals.}

\item{im}{Boolean, set to FALSE by default.  Set to TRUE to calculate CIs following Imbens and Manski (2004)}

\item{k}{Integer, set to 100 by default. Number of bootstrap replications for MIV bias correction.}

\item{ncells}{Integer, set to 5 by default. Number of cells used in the MIV estimator.}
}
\value{
A list of the class 'tebounds'.  Summary and plot methods are provided to match output of the Stata routine.
}
\description{
This function is a port of the Stata command tebounds (Millimet, McCarthy, and Roy, 2015).  tebounds provides estimates of the bounds of the ATE assuming binary treatment (Y) and a binary outcome (D).  The tebounds command
estimates the bounds of the ATE under a range of alternative assumptions regarding the selection process and the extent of measurement error, including assumptions of monotone treatment response and monotone treatment selection.
The command also allows for the inclusion of an MIV.  The monotone treatment response, monotone treatment selection, and MIV approaches follow the estimation procedure developed in Manski and Pepper (2000), Kreider and Pepper (2007), Kreider et al. (2012), and Gundersen, Kreider, and Pepper (2012).
}
\examples{


}
